Добавил проверку функции ini_get_all

+++ admin/phpinfo.php
@@ -29,32 +29,37 @@ if (is_admin(array(101, 102))) {
 	$res = DB::run() -> querySingle("SELECT VERSION()");
 	echo 'PDO MySQL: <b>'.preg_replace('/[^0-9\.]/', '', $res).'</b><br />';
 
-	$ini = ini_get_all();
+	if (function_exists('ini_get_all')) {
+		$ini = ini_get_all();
 
-	echo '<br /><table width="99%" border="0" cellspacing="0" cellpadding="2">';
-	echo '<tr bgcolor="ffff00"><td width="40%">Directive</td><td width="60%">Local Value</td></tr>';
+		echo '<br /><table width="99%" border="0" cellspacing="0" cellpadding="2">';
+		echo '<tr bgcolor="ffff00"><td width="40%">Directive</td><td width="60%">Local Value</td></tr>';
 
-	$q = 0;
-	foreach($ini as $inikey => $inivalue) {
-		$q++;
+		$q = 0;
+		foreach($ini as $inikey => $inivalue) {
+			$q++;
 
-		if ($q&1) {
-			$bgcolor = "#ffffff";
-		} else {
-			$bgcolor = "#e0e0e0";
+			if ($q&1) {
+				$bgcolor = "#ffffff";
+			} else {
+				$bgcolor = "#e0e0e0";
+			}
+
+			if (strlen($inivalue['local_value']) > 40) {
+				$inivalue['local_value'] = substr($inivalue['local_value'], 0, 40);
+				$inivalue['local_value'] .= "...";
+			}
+			if ($inivalue['local_value'] == "") {
+				$inivalue['local_value'] = 'no_value';
+			}
+
+			echo '<tr bgcolor="'.$bgcolor.'"><td width="40%">'.$inikey.'</td><td width="60%">'.check($inivalue['local_value']).'</td></tr>';
 		}
+		echo '</table><br />';
 
-		if (strlen($inivalue['local_value']) > 40) {
-			$inivalue['local_value'] = substr($inivalue['local_value'], 0, 40);
-			$inivalue['local_value'] .= "...";
-		}
-		if ($inivalue['local_value'] == "") {
-			$inivalue['local_value'] = 'no_value';
-		}
-
-		echo '<tr bgcolor="'.$bgcolor.'"><td width="40%">'.$inikey.'</td><td width="60%">'.check($inivalue['local_value']).'</td></tr>';
+	} else {
+		show_error('Ошибка! Функция ini_get_all запрещена в php.ini!');
 	}
-	echo '</table><br />';
 
 	echo '<img src="/images/img/panel.gif" alt="image" /> <a href="index.php">В админку</a><br />';
 } else {
